requirements.txt
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
livekit>=1.0.8
aiortc>=1.5.0
httpx>=0.25.0
tenacity>=8.2.0
redis>=5.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
structlog>=23.2.0
prometheus-client>=0.19.0
uvloop>=0.19.0
websockets>=11.0.2